TIP-1287: move AttributeIsAFamilyVariantAxis in Structure
