+++ src/Akeneo/Pim/Structure/.php_cd.php
@@ -11,7 +11,6 @@ $builder = new RuleBuilder();
 $rules = [
     $builder->only([
         'Doctrine',
-        'Storage',
         'Symfony',
         'Akeneo\Tool',
         'Akeneo\Pim\Structure\Component',
