+++ src/Akeneo/Pim/Structure/Bundle/Query/PublicApi/Attribute/Sql/AttributeIsAFamilyVariantAxis.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Attribute;
+namespace Akeneo\Pim\Structure\Bundle\Query\PublicApi\Attribute\Sql;
 
 use Akeneo\Pim\Structure\Component\Model\VariantAttributeSetInterface;
 use Doctrine\ORM\EntityManagerInterface;
